+++ computer-repair-service/src/main/java/com/solvd/app/dao/mysql/RepairDAO.java
@@ -18,7 +18,7 @@ public class RepairDAO extends MySQLDAO implements IRepairDAO {
     }
 
     @Override
-    public Repair getEntityById(long id) {
+    public Repair getEntityById(int id) {
         return null;
     }
